+++ UtilityBelt/Lib/Networking/Messages/LoginMessage.cs
@@ -1,0 +1,19 @@
+﻿using ProtoBuf;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace UtilityBelt.Lib.Networking.Messages {
+    [ProtoContract]
+    class LoginMessage {
+        [ProtoMember(1)]
+        public string Name { get; set; }
+
+        public LoginMessage() { }
+
+        public LoginMessage(string name) {
+            Name = name;
+        }
+    }
+}