+++ UtilityBelt/Tools/VTankFellowHeals.cs
@@ -6,30 +6,28 @@ using uTank2;
 using static uTank2.PluginCore;
 using Decal.Adapter.Wrappers;
 using System.Runtime.InteropServices;
-using SharedMemory;
 using UtilityBelt.Lib;
 using System.Threading;
 using System.Diagnostics;
 using System.ComponentModel;
 using UBLoader.Lib.Settings;
+using UtilityBelt.Lib.Networking.Messages;
 
 namespace UtilityBelt.Tools {
     [Name("VTankFellowHeals")]
-    [Summary("Automatically forwards vital information to VTank.")]
+    [Summary("Automatically forwards vital and spellcasting information to VTank.")]
     [FullDescription(@"
-If enabled, this will automatically share vital information for all clients on the same PC with VTank.
+If enabled, this will automatically share vital information for all clients on the same PC with VTank.  Spell casting information will also be shared. If you have two characters vulning, they should choose different targets and not overlap spells.
 
 This allows VTank to heal/restam/remana characters on your same pc, even when they do not share in ingame fellowship.
     ")]
     public class VTankFellowHeals : ToolBase {
-        private BufferReadWrite sharedBuffer;
         DateTime lastThought = DateTime.UtcNow;
         DateTime lastUpdate = DateTime.MinValue;
+        private bool isRunning;
 
-        private const string BUFFER_NAME = "UtilityBeltyVTankFellowHealsBuffer";
-        private const int UPDATE_TIMEOUT = 6000; // ms
-        private const int UPDATE_INTERVAL = 500; // ms
-        private int BUFFER_SIZE = 1024 * 1024;
+        public int LastAttemptedSpellId { get; private set; }
+        public int LastAttemptedTarget { get; private set; }
 
         public VTankFellowHeals(UtilityBeltPlugin ub, string name) : base(ub, name) {
 
@@ -37,19 +35,68 @@ This allows VTank to heal/restam/remana characters on your same pc, even when th
 
         public override void Init() {
             base.Init();
-            try {
-                sharedBuffer = new SharedMemory.BufferReadWrite(BUFFER_NAME, BUFFER_SIZE);
 
-                // write a blank record count if we are the first ones here
-                var d = 0;
-                sharedBuffer.Write<int>(ref d);
-            }
-            catch {
-                sharedBuffer = new SharedMemory.BufferReadWrite(BUFFER_NAME);
-            }
+            if (UB.VTank.VitalSharing)
+                TryEnable();
 
+            UB.VTank.VitalSharing.Changed += VitalSharing_Changed;
+        }
+
+        private void VitalSharing_Changed(object sender, SettingChangedEventArgs e) {
+            if (UB.VTank.VitalSharing)
+                TryEnable();
+            else
+                Stop();
+        }
+
+        private void TryEnable() {
+            if (isRunning)
+                return;
             UB.Core.CharacterFilter.ChangeVital += CharacterFilter_ChangeVital;
-            UB.Core.RenderFrame += Core_RenderFrame;
+            UB.Core.EchoFilter.ClientDispatch += EchoFilter_ClientDispatch;
+            UB.Core.ChatBoxMessage += Core_ChatBoxMessage;
+            UBHelper.Core.RadarUpdate += Core_RadarUpdate;
+            UB.Networking.OnPlayerUpdateMessage += Networking_OnPlayerUpdateMessage;
+            UB.Networking.OnCastAttemptMessage += Networking_OnCastAttemptMessage;
+            UB.Networking.OnCastSuccessMessage += Networking_OnCastSuccessMessage;
+            isRunning = true;
+        }
+
+        private void Stop() {
+            if (!isRunning)
+                return;
+            UB.Core.CharacterFilter.ChangeVital -= CharacterFilter_ChangeVital;
+            UB.Core.EchoFilter.ClientDispatch -= EchoFilter_ClientDispatch;
+            UB.Core.ChatBoxMessage -= Core_ChatBoxMessage;
+            UBHelper.Core.RadarUpdate -= Core_RadarUpdate;
+            UB.Networking.OnPlayerUpdateMessage -= Networking_OnPlayerUpdateMessage;
+            UB.Networking.OnCastAttemptMessage -= Networking_OnCastAttemptMessage;
+            UB.Networking.OnCastSuccessMessage -= Networking_OnCastSuccessMessage;
+            isRunning = false;
+        }
+
+        private void Networking_OnPlayerUpdateMessage(object sender, EventArgs e) {
+            if (sender is PlayerUpdateMessage message)
+                UpdateVTankVitalInfo(message);
+        }
+
+        private void Networking_OnCastAttemptMessage(object sender, EventArgs e) {
+            if (sender is CastAttemptMessage message) {
+                UBHelper.vTank.Instance?.LogCastAttempt(message.SpellId, message.Target, message.Skill);
+            }
+        }
+
+        private void Networking_OnCastSuccessMessage(object sender, EventArgs e) {
+            if (sender is CastSuccessMessage message) {
+                UBHelper.vTank.Instance?.LogSpellCast(message.Target, message.SpellId, message.Duration);
+            }
+        }
+
+        private void Core_RadarUpdate(double uptime) {
+            try {
+                UpdateMySharedVitals();
+            }
+            catch (Exception ex) { Logger.LogException(ex); }
         }
 
         private void CharacterFilter_ChangeVital(object sender, ChangeVitalEventArgs e) {
@@ -59,138 +106,73 @@ This allows VTank to heal/restam/remana characters on your same pc, even when th
             catch (Exception ex) { Logger.LogException(ex);  }
         }
 
-        public void UpdateMySharedVitals() {
-            if (UBHelper.vTank.Instance == null || !UB.VTank.VitalSharing) return;
-            UBPlayerUpdate playerUpdate;
+        private void EchoFilter_ClientDispatch(object sender, Decal.Adapter.NetworkMessageEventArgs e) {
             try {
-                playerUpdate = GetMyPlayerUpdate();
-            }
-            catch {
-                return;
-            }
-
-            try {
-                int recordCount = 0;
-                var updates = new List<UBPlayerUpdate>();
-                var i = 0;
-                int offset = sizeof(int);
-
-                using (var mutex = new Mutex(false, "UtilityBelt.VTankFellowHeals.SharedMemory")) {
-                    try {
-                        if (!mutex.WaitOne(TimeSpan.FromMilliseconds(20), true)) {
-                            return;
-                        }
-                    }
-                    catch (AbandonedMutexException) {
-                        return;
-                    }
-
-                    try {
-                        sharedBuffer.Read<int>(out recordCount, 0);
-                        while (i < recordCount && offset <= sharedBuffer.BufferSize) {
-                            UBPlayerUpdate update = new UBPlayerUpdate();
-                            offset = update.Deserialize(sharedBuffer, offset);
-
-                            if (update.PlayerID != UB.Core.CharacterFilter.Id && DateTime.UtcNow - update.lastUpdate <= TimeSpan.FromMilliseconds(UPDATE_TIMEOUT)) {
-                                updates.Add(update);
-                                UpdateVTankVitalInfo(update);
-                            }
-                            else if (update.PlayerID != UB.Core.CharacterFilter.Id) {
-                                //Util.WriteToChat("Marking player as invalid: " + update.PlayerID.ToString() + " on server " + update.Server);
-                                UBHelper.vTank.Instance?.HelperPlayerSetInvalid(update.PlayerID);
-                            }
-
-                            i++;
-                        }
-                    }
-                    catch {
-                        try {
-                            sharedBuffer.Write<int>(ref recordCount, 0);
-                        }
-                        catch { }
-                        return;
-                    }
-
-                    var newRecordCount = updates.Count + 1;
-
-                    //Util.WriteToChat($"Wrote newRecordCount:{newRecordCount} w/ id:{playerUpdate.PlayerID} stam:{playerUpdate.curStam}/{playerUpdate.maxStam}");
-
-                    try {
-                        sharedBuffer.Write(ref newRecordCount, 0);
-                        offset = playerUpdate.Serialize(sharedBuffer, sizeof(int));
-                        for (var x = 0; x < updates.Count; x++) {
-                            offset = updates[x].Serialize(sharedBuffer, offset);
-                        }
-                    }
-                    catch {
-                        newRecordCount = 0;
-                        try {
-                            sharedBuffer.Write<int>(ref recordCount, 0);
-                        }
-                        catch { }
-                    }
-
-                    lastUpdate = DateTime.UtcNow;
+                // Magic_CastTargetedSpell
+                if (e.Message.Type == 0xF7B1 && e.Message.Value<int>("action") == 0x004A) {
+                    var target = e.Message.Value<int>("target");
+                    var spellId = e.Message.Value<int>("spell");
+                    var skill = Spells.GetEffectiveSkillForSpell(spellId);
+                    UB.Networking.SendObject("CastAttemptMessage", new CastAttemptMessage(spellId, target, skill));
+                    LastAttemptedSpellId = spellId;
+                    LastAttemptedTarget = target;
                 }
             }
             catch (Exception ex) { Logger.LogException(ex); }
         }
 
-        private void UpdateVTankVitalInfo(UBPlayerUpdate update) {
+        private void Core_ChatBoxMessage(object sender, Decal.Adapter.ChatTextInterceptEventArgs e) {
+            try {
+                if (LastAttemptedSpellId != 0 && e.Text.StartsWith("You cast ")) {
+                    var duration = Spells.GetSpellDuration(LastAttemptedSpellId) * 1000;
+                    UB.Networking.SendObject("CastSuccessMessage", new CastSuccessMessage(LastAttemptedSpellId, LastAttemptedTarget, duration));
+
+                    LastAttemptedSpellId = 0;
+                }
+            }
+            catch (Exception ex) { Logger.LogException(ex); }
+        }
+
+        public void UpdateMySharedVitals() {
+            if (UBHelper.vTank.Instance == null || !UB.VTank.VitalSharing)
+                return;
+
+            UB.Networking.SendObject("PlayerUpdateMessage", GetMyPlayerUpdate());
+        }
+
+        private void UpdateVTankVitalInfo(PlayerUpdateMessage update) {
             try {
                 if (UBHelper.vTank.Instance == null || update == null) return;
                 if (update.Server != UB.Core.CharacterFilter.Server) return;
 
-                //Util.WriteToChat($"Updating vital info for {update.PlayerID} stam:{update.curStam}/{update.maxStam}");
-
                 var helperUpdate = new sPlayerInfoUpdate() {
                     PlayerID = update.PlayerID,
-                    HasHealthInfo = update.HasHealthInfo,
-                    HasManaInfo = update.HasManaInfo,
-                    HasStamInfo = update.HasStamInfo,
-                    curHealth = update.curHealth,
-                    curMana = update.curMana,
-                    curStam = update.curStam,
-                    maxHealth = update.maxHealth,
-                    maxMana = update.maxMana,
-                    maxStam = update.maxStam
+                    HasHealthInfo = true,
+                    HasManaInfo = true,
+                    HasStamInfo = true,
+                    curHealth = update.CurHealth,
+                    curMana = update.CurMana,
+                    curStam = update.CurStam,
+                    maxHealth = update.MaxHealth,
+                    maxMana = update.MaxMana,
+                    maxStam = update.MaxStam
                 };
-
-                if (helperUpdate != null) {
-                    UBHelper.vTank.Instance.HelperPlayerUpdate(helperUpdate);
-                }
+                UBHelper.vTank.Instance.HelperPlayerUpdate(helperUpdate);
             }
             catch (Exception ex) { Logger.LogException(ex); }
         }
 
-        public void Core_RenderFrame(object sender, EventArgs e) {
-            try {
-                if (DateTime.UtcNow - lastUpdate > TimeSpan.FromMilliseconds(UPDATE_INTERVAL)) {
-                    if (UBHelper.vTank.Instance == null || !UB.VTank.VitalSharing) return;
-
-                    UpdateMySharedVitals();
-                }
-            }
-            catch (Exception ex) { Logger.LogException(ex); }
-        }
-
-        private UBPlayerUpdate GetMyPlayerUpdate() {
-            return new UBPlayerUpdate {
+        private PlayerUpdateMessage GetMyPlayerUpdate() {
+            return new PlayerUpdateMessage {
                 PlayerID = UB.Core.CharacterFilter.Id,
 
-                HasHealthInfo = true,
-                HasManaInfo = true,
-                HasStamInfo = true,
+                CurHealth = UB.Core.CharacterFilter.Vitals[CharFilterVitalType.Health].Current,
+                CurMana = UB.Core.CharacterFilter.Vitals[CharFilterVitalType.Mana].Current,
+                CurStam = UB.Core.CharacterFilter.Vitals[CharFilterVitalType.Stamina].Current,
 
-                curHealth = UB.Core.CharacterFilter.Vitals[CharFilterVitalType.Health].Current,
-                curMana = UB.Core.CharacterFilter.Vitals[CharFilterVitalType.Mana].Current,
-                curStam = UB.Core.CharacterFilter.Vitals[CharFilterVitalType.Stamina].Current,
-
-                maxHealth = UB.Core.CharacterFilter.EffectiveVital[CharFilterVitalType.Health],
-                maxMana = UB.Core.CharacterFilter.EffectiveVital[CharFilterVitalType.Mana],
-                maxStam = UB.Core.CharacterFilter.EffectiveVital[CharFilterVitalType.Stamina],
-
-                lastUpdate = DateTime.UtcNow,
+                MaxHealth = UB.Core.CharacterFilter.EffectiveVital[CharFilterVitalType.Health],
+                MaxMana = UB.Core.CharacterFilter.EffectiveVital[CharFilterVitalType.Mana],
+                MaxStam = UB.Core.CharacterFilter.EffectiveVital[CharFilterVitalType.Stamina],
 
                 Server = UB.Core.CharacterFilter.Server
             };
@@ -200,10 +182,8 @@ This allows VTank to heal/restam/remana characters on your same pc, even when th
         protected override void Dispose(bool disposing) {
             if (!disposedValue) {
                 if (disposing) {
-                    UB.Core.CharacterFilter.ChangeVital -= CharacterFilter_ChangeVital;
-                    UB.Core.RenderFrame -= Core_RenderFrame;
-
-                    sharedBuffer.Dispose();
+                    UB.VTank.VitalSharing.Changed -= VitalSharing_Changed;
+                    Stop();
 
                     base.Dispose(disposing);
                 }